cart_api/requirement.txt
Django>=3.0,<4.0
djangorestframework>=3.12,<4.0
